requirements.txt
# Core dependencies
google-adk
google-genai
nest-asyncio
python-dotenv

# Optional: for local development without Kaggle
jupyter
jupyterlab

